+++ CHANGELOG.md
@@ -16,6 +16,18 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Breaking
 
+## 2025-05-01
+
+- @triton-one/yellowstone-grpc@4.1.0
+- yellowstone-grpc-client-simple-6.1.0
+- yellowstone-grpc-client-6.1.0
+- yellowstone-grpc-geyser-6.1.0
+- yellowstone-grpc-proto-6.1.0
+
+### Features
+
+- proto: add unary method `SubscribeReplayInfo` ([#578](https://github.com/rpcpool/yellowstone-grpc/pull/578))
+
 ## 2025-03-17
 
 - yellowstone-grpc-client-simple-6.0.1